Refactor: Phase 1 (break circular dependency) + Phase 4 (error handling)
Phase 1 — Break circular dependency & clean up types:
- Extract Change, DiffResult, LibraryMetadata, DetectorSpec into new
  checker_types.py module, breaking the checker ↔ suppression circular
  import (was guarded by TYPE_CHECKING)
- Remove unused Detector protocol from detectors.py
- Rename _DetectorSpec → DetectorSpec as the official detector interface
- checker.py re-exports all types for backward compatibility
- suppression.py now imports from checker_types directly

Phase 4 — Standardize error handling:
- Add PolicyError and ReportError to errors.py
- Migrate dumper.py: RuntimeError → SnapshotError (7 sites),
  ValueError → ValidationError (1 site)
- Migrate pdb_parser.py: ValueError → ValidationError (10 sites)
- Migrate package.py: RuntimeError → SnapshotError (9 sites)
- Migrate policy_file.py: ValueError → PolicyError (5 sites)
- Migrate severity.py: ValueError → PolicyError (2 sites)
- Migrate compat/abicc_dump_import.py: ValueError → ValidationError (7 sites)
- Migrate compat/descriptor.py: ValueError → ValidationError (5 sites)
- All new exceptions inherit from both AbicheckError and their original
  builtin type (ValueError/RuntimeError) for backward compatibility
- Move xml.etree.ElementTree.Element import comment in dumper.py to
  clarify it's type-annotation only (parsing uses defusedxml)

All 3464 tests pass with zero test file changes.
